Update IAB Workshop.md
Med and I had an editorial 1:30 session. 

Note that we don't understand: "The NEMOPS workshop intends to focus on managing networks within the control plane (and its effect on the data plane), "
On top of that, this section about "out of scope" should only focus on what IS out of scope.
Shall we keep only the last sentence?